requirements.txt
pymongo
cassandra-driver
redis
requests

pyspark
pandas
numpy
scikit-learn
streamlit
plotly
